+++ src/components/visualizations/chart/chartOptions/test/extendedStackingChartOptions.spec.ts
@@ -1,16 +1,13 @@
 // (C) 2007-2019 GoodData Corporation
 import { Execution } from "@gooddata/typings";
 import { getCategoriesForTwoAttributes } from "../extendedStackingChartOptions";
-import {
-    getDrillableSeries,
-    getSeries,
-    IUnwrappedAttributeHeadersWithItems,
-} from "../../chartOptionsBuilder";
+import { getDrillableSeries, getSeries } from "../../chartOptionsBuilder";
 import { barChartWith4MetricsAndViewBy2Attribute } from "../../../../../../stories/test_data/fixtures";
 import { getMVSForViewByTwoAttributes } from "../../test/helper";
 import { MeasureColorStrategy } from "../../colorFactory";
 import { DEFAULT_COLOR_PALETTE } from "../../../utils/color";
 import * as headerPredicateFactory from "../../../../../factory/HeaderPredicateFactory";
+import { IUnwrappedAttributeHeadersWithItems } from "../../../../../interfaces/ChartOptions";
 
 describe("getCategoriesForTwoAttributes", () => {
     const attributeHeader: Execution.IAttributeHeader["attributeHeader"] = {
